+++ allennlp/commands/train.py
@@ -161,10 +161,7 @@ def train_model(params: Params, serialization_dir: str) -> Model:
                     for instance in dataset
                     if key in datasets_for_vocab_creation)
 
-    total_instances = sum(dataset.num_instances
-                          for key, dataset in all_datasets.items()
-                          if key in datasets_for_vocab_creation)
-    lazy_combined_dataset = LazyDataset(combined_generator, total_instances)
+    lazy_combined_dataset = LazyDataset(combined_generator)
     vocab = Vocabulary.from_params(params.pop("vocabulary", {}), lazy_combined_dataset)
 
     vocab.save_to_files(os.path.join(serialization_dir, "vocabulary"))
