+++ allennlp/data/dataset_readers/conll2003.py
@@ -9,6 +9,7 @@ from allennlp.common import Params
 from allennlp.common.checks import ConfigurationError
 from allennlp.common.file_utils import cached_path
 from allennlp.data.dataset import Dataset
+from allennlp.data.in_memory_dataset import InMemoryDataset
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
 from allennlp.data.fields import TextField, SequenceLabelField
 from allennlp.data.instance import Instance
@@ -118,7 +119,7 @@ class Conll2003DatasetReader(DatasetReader):
         if not instances:
             raise ConfigurationError("reading {} resulted in an empty Dataset".format(file_path))
 
-        return Dataset(instances)
+        return InMemoryDataset(instances)
 
     def text_to_instance(self, tokens: List[Token]) -> Instance:  # type: ignore
         """
