+++ allennlp/data/dataset_readers/seq2seq.py
@@ -8,6 +8,7 @@ import tqdm
 from allennlp.common import Params
 from allennlp.common.checks import ConfigurationError
 from allennlp.data.dataset import Dataset
+from allennlp.data.in_memory_dataset import InMemoryDataset
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
 from allennlp.data.fields import TextField
 from allennlp.data.instance import Instance
@@ -80,7 +81,7 @@ class Seq2SeqDatasetReader(DatasetReader):
                 instances.append(self.text_to_instance(source_sequence, target_sequence))
         if not instances:
             raise ConfigurationError("No instances read!")
-        return Dataset(instances)
+        return InMemoryDataset(instances)
 
     @overrides
     def text_to_instance(self, source_string: str, target_string: str = None) -> Instance:  # type: ignore
