+++ tests/data/dataset_test.py
@@ -5,6 +5,7 @@ import numpy
 from allennlp.common.checks import ConfigurationError
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data import Dataset, Instance, Token, Vocabulary
+from allennlp.data.in_memory_dataset import InMemoryDataset
 from allennlp.data.fields import TextField, LabelField
 from allennlp.data.token_indexers import SingleIdTokenIndexer
 
@@ -24,7 +25,7 @@ class TestDataset(AllenNlpTestCase):
         instance1 = Instance({"tag": (LabelField(1, skip_indexing=True))})
         instance2 = Instance({"words": TextField([Token("hello")], {})})
         with pytest.raises(ConfigurationError):
-            _ = Dataset([instance1, instance2])
+            _ = InMemoryDataset([instance1, instance2])
 
     def test_padding_lengths_uses_max_instance_lengths(self):
         dataset = self.get_dataset()
@@ -81,4 +82,4 @@ class TestDataset(AllenNlpTestCase):
                            self.token_indexer)
         instances = [Instance({"text1": field1, "text2": field2}),
                      Instance({"text1": field3, "text2": field4})]
-        return Dataset(instances)
+        return InMemoryDataset(instances)
