+++ tests/modules/elmo_test.py
@@ -10,6 +10,7 @@ from torch.autograd import Variable
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data.token_indexers.elmo_indexer import ELMoTokenCharactersIndexer
 from allennlp.data import Token, Vocabulary, Dataset, Instance
+from allennlp.data.in_memory_dataset import InMemoryDataset
 from allennlp.data.iterators import BasicIterator
 from allennlp.modules.elmo import _ElmoBiLm, Elmo, _ElmoCharacterEncoder
 from allennlp.data.fields import TextField
@@ -40,7 +41,7 @@ class TestElmoBiLm(AllenNlpTestCase):
                 instance = Instance({"elmo": field})
                 instances.append(instance)
 
-        dataset = Dataset(instances)
+        dataset = InMemoryDataset(instances)
         vocab = Vocabulary()
         dataset.index_instances(vocab)
 
@@ -115,7 +116,7 @@ class TestElmo(AllenNlpTestCase):
             instance = Instance({'elmo': field})
             instances.append(instance)
 
-        dataset = Dataset(instances)
+        dataset = InMemoryDataset(instances)
         vocab = Vocabulary()
         dataset.index_instances(vocab)
         return dataset.as_tensor_dict()['elmo']['character_ids']
